1.换个版本 2.高度雾 3.卡通渲染外描边流程

+++ ShaderLibrary/Input.hlsl
@@ -36,7 +36,7 @@ struct InputData
 {
     float3  positionWS;
     float4  positionCS;
-    float3   normalWS;
+    half3   normalWS;
     half3   viewDirectionWS;
     float4  shadowCoord;
     half    fogCoord;
